allow installed after new installation of Riru
fix RikkaApps/Riru#86

+++ template/magisk_module/riru.sh
@@ -10,13 +10,13 @@ abort_clean() {
 }
 
 check_riru_version() {
-  if [ ! -f "$RIRU_PATH/api_version" ]; then
+  if [ ! -f "$RIRU_PATH/api_version" ] && [ ! -f "$RIRU_PATH/api_version.new" ]; then
     ui_print    "*********************************************************"
     ui_print    "! Riru is not installed"
     ui_print    "! Please install Riru (Riru - Core) from Magisk Manager or https://github.com/RikkaApps/Riru/releases"
     abort_clean "*********************************************************"
   fi
-  RIRU_API_VERSION=$(cat "$RIRU_PATH/api_version") || RIRU_API_VERSION=0
+  RIRU_API_VERSION=$(cat "$RIRU_PATH/api_version.new") || RIRU_API_VERSION=$(cat "$RIRU_PATH/api_version") || RIRU_API_VERSION=0
   [ "$RIRU_API_VERSION" -eq "$RIRU_API_VERSION" ] || RIRU_API_VERSION=0
   ui_print "- Riru API version: $RIRU_API_VERSION"
   if [ "$RIRU_API_VERSION" -lt $RIRU_MIN_API_VERSION ]; then
